requirements.txt
# 머신러닝 및 딥러닝 프레임워크 (안정 버전 - 호환성 테스트 완료)
torch==1.13.1
torchvision==0.14.1
transformers==4.25.1
timm==0.6.12
scikit-learn>=1.3.0
optuna>=3.2.0

# 데이터 처리 및 분석
pandas==2.1.4
numpy==1.26.0
opencv-python-headless==4.8.0.76
albumentations==1.3.1
Pillow==9.4.0

# 시각화
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 실험 관리 및 로깅
wandb>=0.15.0
tensorboard>=2.13.0
mlflow>=2.5.0

# 개발 도구
jupyter==1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
ipykernel==6.27.1
ipython==8.15.0
ipywidgets==8.1.1
matplotlib-inline==0.1.6
tqdm>=4.65.0

# 데이터 검증 및 품질 관리
great-expectations>=0.17.0
pandera>=0.15.0

# 코드 품질
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pytest>=7.4.0

# 유틸리티
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0